requirements.txt
streamlit
pandas
openpyxl
plotly
unidecode
streamlit>=1.30.0
streamlit-extras
gspread
gspread_dataframe
oauth2client
google-api-python-client
requests
cloudinary
streamlit-aggrid
streamlit-searchbox
babel
pandas>=2.2
XlsxWriter>=3.2
openpyxl>=3.1
gspread>=6.0
google-auth>=2.30
gspread-dataframe>=3.3
plotly>=5.23
numpy
google-auth-oauthlib
google-auth-httplib2
yfinance
beautifulsoup4
requests
lxml